Move commondata to its own module
